**Merge a 3ce26c2541dcf9d1f01642289acb4495586b50b0 commit from
watabou/pixel-dungeon to 1.7.2 version.
**Translation of all texts and information concerning the Missiles, on
Classes in "\weapon\missiles" package.
**(sorry, I should have separated the Changes of the Merge)

+++ src/com/watabou/pixeldungeon/levels/features/HighGrass.java
@@ -17,6 +17,7 @@
  */
 package com.watabou.pixeldungeon.levels.features;
 
+import com.watabou.pixeldungeon.Challenges;
 import com.watabou.pixeldungeon.Dungeon;
 import com.watabou.pixeldungeon.actors.Char;
 import com.watabou.pixeldungeon.actors.buffs.Barkskin;
@@ -40,22 +41,24 @@ public class HighGrass {
 		Level.set( pos, Terrain.GRASS );
 		GameScene.updateMap( pos );
 		
-		int herbalismLevel = 0;
-		if (ch != null) {
-			Herbalism herbalism = ch.buff( Herbalism.class );
-			if (herbalism != null) {
-				herbalismLevel = herbalism.level;
+		if (!Dungeon.isChallenged( Challenges.NO_HERBALISM )) {
+			int herbalismLevel = 0;
+			if (ch != null) {
+				Herbalism herbalism = ch.buff( Herbalism.class );
+				if (herbalism != null) {
+					herbalismLevel = herbalism.level;
+				}
 			}
-		}
-		
-		// Seed
-		if (herbalismLevel >= 0 && Random.Int( 18 ) <= Random.Int( herbalismLevel + 1 )) {
-			level.drop( Generator.random( Generator.Category.SEED ), pos ).sprite.drop();
-		}
-		
-		// Dew
-		if (herbalismLevel >= 0 && Random.Int( 6 ) <= Random.Int( herbalismLevel + 1 )) {
-			level.drop( new Dewdrop(), pos ).sprite.drop();
+			
+			// Seed
+			if (herbalismLevel >= 0 && Random.Int( 18 ) <= Random.Int( herbalismLevel + 1 )) {
+				level.drop( Generator.random( Generator.Category.SEED ), pos ).sprite.drop();
+			}
+			
+			// Dew
+			if (herbalismLevel >= 0 && Random.Int( 6 ) <= Random.Int( herbalismLevel + 1 )) {
+				level.drop( new Dewdrop(), pos ).sprite.drop();
+			}
 		}
 		
 		int leaves = 4;
